fix(bug): crashing when trying to remove alot of files/directories
That's because resetting the allocator position in a recursive function
cause problems. The allocator was resetting the correct memory but some
memory was used by previous calls? I donno for sure. To avoid this
problem I simply created a new 'Temp_String' strcut that has a
destrcutor and calls 'allocator.reset' with the capacity of the
allocated string.