Arena remove 'alloc' optional 'context' argument
That's because the context was mainly begin used in the 'New_String'
type in order to produce an 'Allocator' interface that would have the
pointer to the arena in which the string buffer is allocated.

We simply don't do that anymore and instead only have the pointer to the 'Arena'
and it takes care-of allocating space for extra memory for the string buffer.